added demolish visualization

+++ src/app/WIV/wiv-viewer/wiv-viewer.component.ts
@@ -5,7 +5,7 @@ import { PropertiesService } from '../wiv-services/properties.service';
 import { activityType, ContextMenuButtons, ElementClickEvent, ModelViewerSettings, PickedObject, propType, schedulePropClass, SimulationEl, XYScreenCoordinate } from '../models';
 import { IfcViewerAPI } from 'web-ifc-viewer';
 import { DatesRange, RangeSliderService } from '../wiv-services/slider.service';
-import { Color, DoubleSide, MeshLambertMaterial } from 'three';
+import { Color, DoubleSide, Material, MeshLambertMaterial } from 'three';
 
 @Component({
   selector: 'app-wiv-viewer',
@@ -23,6 +23,7 @@ export class WivViewerComponent implements OnInit {
   public contextMenuPosition?: XYScreenCoordinate;
   public elementClickContext?: ElementClickEvent;
   public pickingMode?: boolean;
+  public simulationMode: boolean = false;
 
   // Models
   public models: any = [];
@@ -31,6 +32,13 @@ export class WivViewerComponent implements OnInit {
   fileName = '';
   public model: any;
   public modelId?: number;
+
+  // Mode
+  public baselineDates: boolean = false;
+  public startDateProp: string = "PAA_4D_Actual Start";
+  public finishDateProp: string = "PAA_4D_Actual Finish";
+  public demStartDateProp: string = "PAA_4D_Actual Start Demolish";
+  public demFinishDateProp: string = "PAA_4D_Actual Finish Demolish";
 
   // Properties
   public contextMenuButtons: ContextMenuButtons
@@ -54,7 +62,17 @@ export class WivViewerComponent implements OnInit {
   public currentElements: any[] = [];
   public existingElements: any[] = [];
   public hiddenElements: any[] = [];
+  public currentDemolishedElements: any[] = [];
+  public demolishedElements: any[] = [];
 
+  public existingElementssubset: any;
+  public highlightedElementsSubset: any;
+  public demolishElementsSubset: any;
+
+  public loadedSubsets: string[] = [];            // List of customIDs of subsets that are loaded in the scene
+  public loadedSubsetMaterials: Material[] = [];  // List of materials of subsets that are loaded in the scene
+  public loadedSubsetExpressIDs: any[] = [];      // List of expressIDs of subsets that are loaded in the scene
+  private subsetModelId: any = {}; // Main with subset name as key and modelID as Value
 
   public scheduleProperties = [
     {expressID: undefined, Name: "PAA_4D-Name", value: undefined, type: propType.STRING, activityType: activityType.CONSTRUCT},
@@ -70,6 +88,8 @@ export class WivViewerComponent implements OnInit {
     {expressID: undefined, Name: "PAA_4D_Baseline Start Demolish", value: undefined, type: propType.DATE, activityType: activityType.DEMOLISH},
     {expressID: undefined, Name: "PAA_4D_Baseline Finish Demolish", value: undefined, type: propType.DATE, activityType: activityType.DEMOLISH},
   ]
+
+  public currentActivities: any[] = []
 
   constructor(
     private _viewer: WebIfcViewerService,
@@ -100,7 +120,11 @@ export class WivViewerComponent implements OnInit {
     }
 
 
-    this.pickingMode = true
+    if(this.simulationMode) {
+      await this.initSimulationFromProps()
+    } else {
+      this.pickingMode = true
+    }
 
     // Set viewer behavior
     // const el = this.highlightSelection(0, this.selection, true);
@@ -166,6 +190,9 @@ export class WivViewerComponent implements OnInit {
 
   public async initSimulationFromProps() {
     this.pickingMode = false;
+    this.simulationMode = true;
+
+    console.log(this.baselineDates)
 
     if(!this.viewer) return
     for (let i = 0; i < this.models.length; i++) {
@@ -185,11 +212,23 @@ export class WivViewerComponent implements OnInit {
       this.dates = [...this.dates, ...dates];
       this.dates.sort((date1, date2) => date1.valueOf() - date2.valueOf());
     }
-
     console.log(this.simulationList)
 
-    this.startDate = this.dates[0];
-    this.endDate = this.dates.slice(-1)[0];
+    // this.startDate = this.dates[0];
+    // this.endDate = this.dates.slice(-1)[0];
+    let startDate = new Date();
+    let endDate = new Date();
+
+    // 14 days earlier
+    var dateOffset = (24*60*60*1000) * 14; //5 days
+    let startDateNumber = JSON.parse(JSON.stringify(this.dates[0].getTime()));
+    startDate.setTime(startDateNumber - dateOffset);
+    this.startDate = startDate
+
+    let endDateNumber = JSON.parse(JSON.stringify(this.dates.slice(-1)[0].getTime()));
+    endDate.setTime(endDateNumber + dateOffset);
+    this.endDate = endDate;
+
     console.log(this.startDate, this.endDate)
 
     if(this.endDate && this.startDate) {
@@ -200,11 +239,16 @@ export class WivViewerComponent implements OnInit {
       console.log("dates are not correct")
     }
 
+    this.dimModelsAndRemoveShadow()
+  }
+
+  dimModelsAndRemoveShadow() {
+    if(!this.viewer) return
     // Hide model, so subsets can be shown
     for (let i = 0; i < this.models.length; i++) {
       // this.hideModel(0);
       this.dimModel(i);
-      this.viewer.shadowDropper.deleteShadow(i.toString());
+      if(this.viewer.shadowDropper.shadows[i.toString()]) this.viewer.shadowDropper.deleteShadow(i.toString());
     }
   }
 
@@ -218,25 +262,37 @@ export class WivViewerComponent implements OnInit {
 
     const scene = this.viewer.context.getScene();
 
+    if(this.baselineDates == false) {
+      this.startDateProp = "PAA_4D_Actual Start";
+      this.finishDateProp = "PAA_4D_Actual Finish";
+      this.demStartDateProp = "PAA_4D_Actual Start Demolish";
+      this.demFinishDateProp = "PAA_4D_Actual Finish Demolish";
+    } else {
+      this.startDateProp = "PAA_4D_Baseline Start";
+      this.finishDateProp = "PAA_4D_Baseline Finish";
+      this.demStartDateProp = "PAA_4D_Baseline Start Demolish";
+      this.demFinishDateProp = "PAA_4D_Baseline Finish Demolish";
+    }
+
     // To DO:
     // Improve to only loop over models once --> Strucutre simulationList also by models and then create subsets for each model after and add it to scene within the loop
 
     // create empty arrays for each model for current and existing elements --> Set empty when slider changes!
     this.currentElements = Array(this.models.length).fill([]);
+    this.currentDemolishedElements = Array(this.models.length).fill([]);
+    this.demolishedElements = Array(this.models.length).fill([]);
+    this.currentElements = Array(this.models.length).fill([]);
     this.existingElements = Array(this.models.length).fill([]);
 
     for (let i = 0; i < this.models.length; i++) {
 
-      // Dim model
-      // this.dimModel(i);
-
-      // To DO
-      // To DO
-      // To DO
-      // Rather check current and new elements from before and from now and remove or add from subset!
-
-      let currentElements = [];
-      let existingElements = [];
+      let exName = JSON.stringify(i) + "_existing elements";
+      let curName = JSON.stringify(i) + "_current elements";
+      let demName = JSON.stringify(i) + "_demolished elements";
+      let currentElements: any[] = [];
+      let currentDemolishedElements: any[] = [];
+      let demolishedElements: any[] = [];
+      let existingElements: any[] = [];
 
       const simulationList = this.simulationList[i];
       for (let j = 0; j < simulationList.length; j++) {
@@ -244,10 +300,18 @@ export class WivViewerComponent implements OnInit {
         const modelID = item.modelID;
         const scheduleProps = item.scheduleProperties as schedulePropClass[];
 
-        const startDate = scheduleProps?.filter((item:any) => item.Name === 'PAA_4D_Actual Start')[0].value as Date;
-        const finishDate = scheduleProps?.filter((item:any) => item.Name === 'PAA_4D_Actual Finish')[0].value as Date;
+
+        const startDate = scheduleProps?.filter((item:any) => item.Name === this.startDateProp)[0].value as Date;
+        const finishDate = scheduleProps?.filter((item:any) => item.Name === this.finishDateProp)[0].value as Date;
+
+        const demStartDateList = scheduleProps?.filter((item:any) => item.Name === this.demStartDateProp);
+        const demFinishDateList = scheduleProps?.filter((item:any) => item.Name === this.demFinishDateProp);
+        let demStartDate, demFinishDate;
+        demStartDateList.length > 0 ? demStartDate = demStartDateList[0].value as Date : demStartDate = undefined;
+        demFinishDateList.length > 0 ? demFinishDate = demFinishDateList[0].value as Date : demFinishDate = undefined;
         // console.log(startDate, finishDate, currentDate)
 
+        // CONSTRUCTION
         if (startDate <= currentDate && currentDate <= finishDate) {
           // currently built
           currentElements.push(item.expressID);
@@ -262,21 +326,33 @@ export class WivViewerComponent implements OnInit {
           // built in future
           // this.hiddenElements.push(item.expressID);
         }
+
+        // DEMOLISHING
+        if (demStartDate && demFinishDate) {
+          if (demStartDate <= currentDate && currentDate <= demFinishDate) {
+            // currently demolished
+            currentDemolishedElements.push(item.expressID);
+            this.currentDemolishedElements[modelID].push(item.expressID);
+            // console.log(currentDate, item, this.currentElements, this.currentElements[modelID])
+          } else if (demFinishDate < currentDate) {
+            // already demolished --> Update existing elements
+            existingElements = existingElements.filter(x => x != item.expressID)
+            demolishedElements.push(item.expressID);
+            // console.log(demolishedElements)
+            this.demolishedElements[modelID].push(item.expressID);
+          }
+        }
       }
 
       const modelID = i;
-      // Show existing with grey color?
-      let exName = JSON.stringify(i) + "_existing elements";
-      const existingElementssubset = this.createSubsetOfIds(modelID, existingElements, exName, "#fcfcfc");
-      // console.log(existingElementssubset)
-      if(existingElementssubset) scene.add(existingElementssubset);
+      // EXISTING
+      this.existingElementssubset = this.createSubsetOfIds(modelID, existingElements, exName, "#dbdbdb");
 
-      // highlight
-      let curName = JSON.stringify(i) + "_current elements";
-      // const highlightedElementsSubset = this.createSubsetOfIds(modelID, currentElements, curName, "yellow", 0.5);
-      const highlightedElementsSubset = this.createSubsetOfIds(modelID, currentElements, curName, "yellow");
-      // console.log(highlightedElementsSubset)
-      if(highlightedElementsSubset) scene.add(highlightedElementsSubset);
+      // NEW
+      this.highlightedElementsSubset = this.createSubsetOfIds(modelID, currentElements, curName, "yellow");
+
+      // CURRENT DEMOLISHED
+      this.demolishElementsSubset = this.createSubsetOfIds(modelID, currentDemolishedElements, demName, "red");
     }
   }
 
@@ -287,8 +363,18 @@ export class WivViewerComponent implements OnInit {
     return viewerApi.IFC.selector.highlight.newSelection(modelID, ids, removePrevious)
   }
 
-  createSubsetOfIds(modelID: number, ids: number[], name: string, color?: string, opacity?: number) {
+  createSubsetOfIds(modelID: number, ids: number[], customID: string, color?: string, opacity?: number) {
     if(!this.viewer) return
+
+    // Update model ID map
+    this.subsetModelId[customID] = modelID;
+
+    // If a subset with this id is loaded already, start by removing it
+    const subsetIndex = this.loadedSubsets.indexOf(customID);
+    if(subsetIndex != -1){
+        // console.log("remove subset in color service: ", customID)
+        this.removeSubset(customID);
+    }
 
     let material = undefined
     if(color != undefined) {
@@ -303,7 +389,7 @@ export class WivViewerComponent implements OnInit {
       }
     }
 
-
+    // Create new subset
     const scene = this.viewer.context.getScene()
     const subset = this.viewer.IFC.loader.ifcManager.createSubset({
       modelID,
@@ -311,9 +397,48 @@ export class WivViewerComponent implements OnInit {
       ids,
       removePrevious: true,
       material,
-      customID: name,
+      customID,
     });
+    if(subset.name == "") subset.name = customID;
+
+    this.loadedSubsets.push(customID);
+    this.loadedSubsetMaterials.push(material!);
+    this.loadedSubsetExpressIDs.push(ids);
+
     return subset
+  }
+
+  public removeFromSubset(modelID: number, ids: number[], customID: string) {
+    if(!this.viewer) return;
+
+    // If a subset with this id is loaded, remove items
+    const subsetIndex = this.loadedSubsets.indexOf(customID);
+    if(subsetIndex != -1){
+        // console.log("remove subset in color service: ", customID)
+        this.viewer.IFC.loader.ifcManager.removeFromSubset(modelID, ids, customID)
+    }
+  }
+
+  public removeSubset(customID: string){
+
+    if(!this.viewer) return;
+
+    // Get modelID
+    const modelID = this.subsetModelId[customID];
+
+    const subsetIndex = this.loadedSubsets.indexOf(customID);
+    if(subsetIndex != -1){
+
+      // Get the material that was appended to the subset
+      const subsetMaterial = this.loadedSubsetMaterials[subsetIndex];
+      const expressIDs = this.loadedSubsetExpressIDs[subsetIndex];
+
+      this.viewer.IFC.loader.ifcManager.removeSubset(modelID, subsetMaterial, customID);
+      this.loadedSubsets.splice(subsetIndex, 1);
+      this.loadedSubsetMaterials.splice(subsetIndex, 1);
+
+    }
+
   }
 
   public hideModel(modelID: number){
